requirements.txt
pytest>=4.6
pytest-cov>=2.6.1
pytest-xdist>=1.25.0
coveralls>=1.5.1
antlr4-python3-runtime>=4.9.1
dataclasses
